Adjust chromereload.js to do a smart reload of extension pages
Reload requests are debounced over 300ms to avoid repeated reloads, and
then rather than reloading the whole extension or background page, the
path whose reload is requested is taken into account and only that page
is reloaded. When the popup page has changes, this means the popup page
will be reloaded independent of the rest of the extension, making for a
smoother reload experience.