Add WebExtesionArchivePlugin and integrate for prod builds
The plugin is in build-utils and is type-checked with TypeScript since
it interacts with Webpack internals.

The webpack config has been adjusted so that mode-specific
configurations are functions that take a browser and produce
configuration adjustments based on the browser. Currently this is only
used for production adjustments.

Finally, the webpack config is now also set to be type-checked, though
it does not fully pass those checks yet.