Adjust README introductory information
The biggest adjustment is an expanded "Building and Developing" section
with a new quickstart subsection that links to instructions for getting
a dev build up and running on a local browser.

Also clarifies that the yarn build command will produce ZIPs in the dist
directory (though this branch does not yet do so).