Fix WebSocket endpoint check
As articulated, an error would be thrown if the endpoint didn't have
EITHER wss or ws://, but of course wss:// doesn't include ws://, so it
always failed even for WebSockets.

Switch to a regex to avoid the double-negative check that is more
complicated to reason about. ALso anchor the match to the start of the
string to validate the endpoint more thoroughly.