Real-time Ethermine data for account subscriptions
Rather than stand up the full account infrastructure, the code switches
out the subscription stub to instead connect to Blocknative and look for
events related to the Ethermine account, which is very active in
ETH interactions. This gives a good stream of data to work with.

Notably, it also means very high event usage <_<