Periodically pull token balances from Alchemy
Any accounts and tokens flagged for tracking will get periodically have
their balances checked. No accounts of tokens are added to the tracking
list by default — that can be handled in future work as the backend
state is better unified.

As it stands, this requires a fair amount of Alchemy API calls. This can
be optimized in the future via the strategies described in #136

Closes #125, refs #136