Move api Main class to its own file
Also change the return type of `startApi` to a bare `Main` instance. The
container object for API info seems like it's settling out to `Main`
rather than something that includes it, and can be adjusted back if
needed. Moreover, currently no external code is interacting with the
returned value anymore.