Validate Alchemy assetTransfer API responses
Good to establish a ground truth here around third-party API details.
Many won't maintain strong versioning, and having solid validation means
we can point out when they haven't maintained backwards compatibility...
as well as help prevent potential attacks from malicious providers. Any
malformed asset transfers here are dropped, logging a warning to the
console. In the future, we should be considerate about when we want to
drop and log, when to throw, and whether it's ever appropriate to allow
strictly invalid data through to lessen the potential impact of an
"outage" from an API that's gone through an errant upgrade.

In this particular case, of course, Alchemy has been great... and I'm
the one who needs a little more structure 🤪

This validation step was done with Ajv and JSON Type Definitions. Even
though they're newer than JSON Schemas, I believe JTD is going to be a
more ergonomic solution for us for most things.

More details can be found in RFC 8927 or at https://jsontypedef.com/.