Fix primary key types for chain db
These types were declared as `number`, but were actually string arrays.
TypeScript did not catch the usage of `.get` with the wrong type due to
the way that Dexie declares overloads of the `get` method; a bug has
been filed with Dexie to try to improve this in the future.