Fix IndexedDB transaction issues in chain db
Due to the way that IndexedDB transactions work, using an async function
in the transaction can lead to db transactions that are auto-committed
before the actual transaction code finishes.

addOrUpdateTransaction was using a read-write transaction to do a get-
or-else-update piece of logic that only updated certain fields if
a chain transaction was already in the db. This required doing an async
read of the database, which triggered the above condition.

To fix this, the code now always uses a `.put`, which is always a
get-or-else-update operation. When the database is created, an
`updating` hook is installed to limit which chain transaction properties
can be updated on an existing chain transaction and how.