Make Dexie tables private in service databases
Direct calls to the tables are now wrapped, per the Law of Demeter.
Migrations are handled in an ugly but safe way that needs further
consideration, and arbitrarily new instances of the databases cannot be
externally created; only getOrCreateDB is exposed to other modules.

For the chain database, blocks are also now `put` instead of `add`ed.
Currently this doesn't expose any information on reorgs, but could be a
good starting point for bubbling this information out.