Add a createBackgroundAsyncThunk helper for redux
This helper is similar to redux toolkit's createAsyncThunk, but it
ensures that the async thunk is handled in the background script no
matter where the action originates.

To do this, a central variable tracks all actions that are created for
consumption in the background script, and the background store
initializer uses the webext-redux `alias` middleware to set up those
actions for background processing. The alias middleware is installed
first so that the async thunk action can be processed through the async
thunk middleware once it hits the background script.