Fix missing cache during redux initialization
If REDUX_CACHE was set to "true", but there was no saved state (say,
because this is the first execution of the extension in this state), the
initialization code would attempt to parse an `undefined`, which throws
(this is standard JSON.parse behavior, unfortunately). State existence
is now checked before attempting to parse the saved state.