Simplify JSON serialization tests
Digging much more deeply in here, the particular issue that was
manifesting with json-bigint looked like it was due to a parse issue
with a large JSON blob, but actually it was due to a very large BigInt
serializing as a regular number due to internal logic issues in the
library.

The updated tests check for large and small bigints and successful
round-tripping both standalone and with other data types. Other than
incidentally, they do not check standard JSON serialization beyond the
minimum necessary.