Component props are no longer optional when defaults are provided
To do this, disable some subparts of two linter rules that required the
component prop types to be optional if they were mentioned in
defaultProps. In older versions of TypeScript, declaring a prop required
when there was a default provided did not allow callers to omit the
prop correctly when using the component. More recent versions of
TypeScript (3+) now correctly expose an optional property externally
while treating it as required inside the body of the component.

All defaultProps-mentioned properties are now marked required in prop
types, laying the groundwork for enabling strict mode.