Make service types strict-mode compatible
This involved two changes: Emittery's primary type includes a `debug`
property that is locked to the event types in strict mode, so the
adjusted Emittery type used in the Service interface now omits that
property.

Additionally, the started() type was using an implicitly any `this` type
that wasn't flagged by noImplicitAny but is flagged by noImplicitThis
when strict mode is flipped on. The type now correctly references the
type of the containing class.