Add AnyEVMBlock and use it in most places
AnyEVMBlock represents the fact that not all blocks are EIP1559 blocks.
Existing code assumed this was true, something which never errored as
strict null checks were required to detect the issues with it. Strict
mode will make these errors, and AnyEVMBlock improves expectations
across the board to match this correct behavior.