Improve and strictify transaction type hierarchy
At the type level, the big differences are the optionalization of the
`to` field (which is not specified in contract-creating transactions),
the addition of the gasLimit field (which exists for all transaction
types, including those that are confirmed), the conversion of all
transaction types from `interface` to `type` declarations, and the
renaming of the `gas` field to `gasUsed` for clarity. This last one is a
mismatch to the default JSON-RPC name, but is clearer for internal use.

Additionally, the concept of an almost-signed-confirmed transaction vs
a confirmed transaction, which existed solely due to an expectation that
confirmed transactions had no gas limit, is gone. In its place is the
concept of an almost-signed transaction, whose signature fields are
optional.

Last but not least, a variety of places, especially in
external-to-internal type conversion, were refined to reflect proper
null/undefined type handling, and to reflect updated field names and
types.