Adjust price validation for strict mode
To do this, a new method jsonSchemaValidatorFor is introduced. It works
the same way as jtdValidatorFor, but uses JSON Schema. The Coingecko
price JTD schema is adapted to a JSON schema that correctly captures the
underlying types, and the prices.ts consumers properly handle the
possibility of an undefined price for a given asset pair.

The Coingecko price schema is odd in that it has a field that is a
number (last_updated_at) alongside an arbitrary number of string keys
that are either a number or undefined. Due to a bug in Ajv's typing, the
JSON schema needs a little massaging in the type system to properly fit
into the expectations; this is detailed alongside the schema in a
comment.