Rework redux initialization for strict initialization requirements
For strict initialization, all fields must either be declared as
possibly-null or they must be assigned directly in the constructor. To
do this, the saved redux state is now resolved in the `create`
function (since it requires a promise) and the store is initialized and
wrapped directly in the Main constructor.

The `initializeRedux` method remains, but simply acts as a kickoff point
to connect the various services to the redux store.