Improve setShowingActivityDetail typing in UI slice
Leaving types off as was previously happening leads to `any` everywhere,
but evidently not in a way that noImplicitAny catches. This is a pitfall
that probably needs to be addressed systemically, but for now improve
the typing---and fix a related bug where setShowingActivityDetail was
being invoked with `undefined` and breaking the activity detail view.