Add update-version script for branch management
When bumping the release version using yarn version, this script checks
whether the branch matches the expected format of release branch
(release-$version). If it does not, it attempts to branch orign/main to
a branch with that name.

Once all this is done, it bumps the manifest version and stages the
manifest as well so that this bump is included in the version commit.