add more background protection
window.postMessage is a broadcast channel and anybody can listen to any port if knows the name.

 So I used some built in browser apis to gather generated extension specific information.

Then I collected these info on every stage (inpage, content, background) and compared them against each other. This way it's almost entirely sure that the message comes from the source that we trust.