Remove assets/networks circular dependency
Networks defines its own NetworkBaseAsset which is structurally
compatible with FungibleAsset. Whether this is better than the circular
dependency or not is in the eye of the beholder, but TypeScript ensures
that these types should be interchangeable unless someone starts
changing them to be structurally incompatible with each other---in which
case the compiler should get mad if it would break stuff.