Factor out account user value enrichment to selector
Until now, this was being partially done in the reducer and partially in
the selector. The selector overwrote things that the reducer had already
done, and the code paths were entirely separate.

Here, everything the reducer was doing is stripped out completely,
leaving the selector in charge of doing all enrichment of the asset
amounts for UI display.