Refactor target strings into constants for communication.
I went with only putting these constants into a variable and using them in place.

When I wanted to export/import them I had the dependency issue, because this single constant import would include the whole bundle.

A solution for this would be to create a `provider-bridge-shared` package and a `constants.ts` in that package so only those things gets shared that we intend to share. But that felt like a heavy touch for this problem so I kept it simple (and a bit redundant) for now.