Fill in fixed point convenience library
The library now supports a paired amount/decimals type called
FixedPointNumber that can be straightforwardly JSON-serialized using
our bigint augmentation, and can be used to track fixed point decimal
changes through operations. Additionally, it supports parsing a standard
US decimal string (potentially with thousands separators) to a
FixedPointNumber.

The external API surface is unchanged, but the functions are written to
work with each other aggressively to minimize the opportunity for error.