Add selectAssetPricePoint to assets slice
This selector allows for the lookup of a particular price point between
two assets, in a known order, for usage by a caller. It takes care of
flipping asset order in the price point if needed, as well as handling
the asset search process.

Currently, it expects a direct price between two assets to return a
price point, but the abstraction will allow for price points that need
to convert through an intermediary asset if desired.