Fix top menu switcher React nested button warning
The protocol switcher was unnecessarily made up of two nested buttons
with only one click handler. Additionally, a button was being used
higher up still to catch the hidden dev menu actions.

The switcher nesting is fixed by simply removing one of the buttons and
adjusting the placement of the click handler accordingly. The hidden dev
menu is now handled by a the same click handler, but the containing
element is a div instead of a button. The reasoning here is that the dev
menu doesn't actually come up during button behavior, which is to take
an action on click, but rather during custom behavior (triple-clicking).
This reasoning is reflected in the skipping of accessibility lint rules.