Add a runAtStart property to alarms and use it
This optional property allows a service to configure an alarm that
should, in addition to firing periodically or in the future, run once
at service start. This speeds up certain initial lookups like the prices
of core cryptocurrencies like BTC and ETH, as well as the first checks
for asset transfers and user-owned token balances.