Add accountSelectors and reexported selectors/ package
accountSelectors now holds selectAccountAndTimestampedActivities, but
just as important selectors are imported directly from the top-level
`redux-slices/selectors` package instead of going to the particular
subpackage for simplicity.

In the process, use a RootState defined at the top level of redux-slices
that can be used in the selectors package, since there is no longer a
circular dependencies concern.