Populate account selector with real data
In addition to the already-functioning address and ENS name in the
list, this commit adds the populating of correct per-address totals,
updates the account grouping to indicate that the addresses are
read-only, and replaces the bolded name with the shortened address in
cases where there is no ENS name to work with. It also populates the
correct avatar, when available.

A new selectAccountTotals selector is introduced that does per-account
totaling to make this possible. There is likely an opportunity for
refactoring of some of the code that is similar with
selectAccountAndTimestampedActivities.