Correctly populate unit price for overview table
When enrichAssetAmountWithMainCurrencyValues was implemented, it lost
the information on unit price for assets. This was originally named
`pricePerToken` and `localizedPricePerToken`, but these terms assume
the asset is a token, which not all assets are.

The fields are now unitPrice and localizedUnitPrice, and
enrichAssetAmountWithMainCurrencyValues now correctly populates these,
and the overview assets table correctly displays them.