Add useIsKeyringUnlocked and use in SignTransaction
This hook checks if the keyring is unlocked, and offers the possibility
of redirecting to an unlock page if needed. If the keyring is currently
locked and uninitialized, the redirect will be to a page that will allow
password creation; otherwise, it will be to a page that will unlock the
keyring. An unlocked keyring will never redirect.