Rework Import Metamask page to use new keyring behavior
In particular, use `useAreKeyringsUnlocked` to ensure the keyrings are
unlocked before allowing import and before redirecting to the next page.
Amongst other things, this ensures that when Metamask import begins, if
the keyrings need a password created or need to be unlocked, that will
be handled transparently.

Also move from an `onImported` event handler with arbitrary behavior to
an explicit `nextPage` prop that redirects to a page once the import is
complete.