Handle undefineds in selectCurrentAccountActivitiesWithTimestamps
The way the function was written, typing indicated there was no
possibility of an undefined activity list; however, in practice, if for
some reason there is no current account, the list was indeed undefined.

selectCurrentAccountActivitiesWithTimestamps is now explicitly written
to have a | undefined type, and the two consumers of it correctly sub in
an empty list if there is no current account activity available.