Don't clear redux keyrings metadata on lock
Due to how the keyring service works, when keyrings are locked, it emits
a keyring update clearing the list of available keyrings. This is
strictly correct, but the keyrings slice in redux should maintain its
metadata about available keyrings so that the UI can still distinguish
between addresses that have signing available (after unlocking) and
those that don't.

To address, this, the `updateKeyrings` reducer now ignores zero-length
keyring list updates if the keyrings are locked.