Change var names and outsource fee button to comp
Some variables were using the word gas while they were referring
to fee, they have been renamed. Saw in new mock that the
button responsible for opening gas option should be on the
Sign page, since it is going to be reused it made sense to
move it out to separate component. Another change in this
commit that is a fix of persisting user fee type choice
when new fees load.