Permissions are persisted into indexedDb
Removed the this.#allowedPages from the `provider-bridge.ts` service to make the db as the source of truth.

Created a mirror of the allowedPages list in the `dapp-permisson` redux slice and update it through the reducers. This can cause problems, because now it's using optimistic updates — meaning it assumes that the data will be persisted in the db — without implemented rollback or safety checks.

This optimistic update operation seemed to be a reasonable assumption now, but it's possible that in the future this has to be rewritten in a way that the indexedDb is the only source of truth and the data is read from it. But for now it seemed unnecessary.