Clarify requirements for transaction request estimator
Gas limit, max fee, value, etc, are all optional when in the process of
constructing a transaction; it's only when actually trying to submit the
request to the chain that they are required. When estimating gas,
default everything to 0x0 except the sender address, which must be set.

Additionally, rather than controlling for `from` conditionally, enforce
it at the type level so that updateTransactionOptions can only run
when `from` is known to be set.