Populate default names and avatars for new accounts
These are pulled in such a way as to avoid reusing unless all names are
used once already. Notably, an account that pulls in an ENS name and
avatar after loading will still consume an default avatar and name,
otherwise the asynchronous nature would force other already allocated
names after it to change.