Fix default account name to be unique until it wraps
The existing approach was choosing from an ever-smaller set of available
name indices, but the names themselves were not being rotated out of the
running. This now happens.

Additionally, the number of existing accounts was including the current
account, which reduced the number of available names by one. This is
also fixed now.