Manage transaction auto-broadcasting in the UI
transaction-construction gets two new fields, one to track the signed
transaction and one to track whether it should be broadcast once signed.
The Send page sends this to true, while the wiring for the internal
Ethereum provider sets it to false. The internal Ethereum provider
manages its own broadcasts, and can pass on broadcasting for certain
calls.

Finally, the SignTransaction page dispatches the broadcast if
broadcastOnSign is set to true at signing time.

This fixes an issue where dApp signing would broadcast a transaction
twice, which amongst other things would cause an exception to interrupt
the process of reporting the successful transaction back to the dApp.