Support multiple networks in ChainService
The goal here is to parameterize the network, *everywhere*... removing
any system-wide network state, and only special-casing Ethereum where
absolutely necessary. For now, that means assuming Ethereum is the first
chain at startup.

It's already become clear that we need a way to be able to test the
capabilities of a particular provider. Alchemy doesn't support most of
its enhanced APIs across all chains, and won't be our only provider for
much longer. Here, I've settled on some very loose type guards to solve
the issue, but as we get more providers and local nodes involved, that
won't cut it.

Continuing our march toward testnet and L2 support.

Refs #5, refs #7