Sign transaction shows sender account in top bar
To do this, SharedCurrentAccountInformation was refactored to take the
account information being displayed from the caller. The top menu
profile button's use passes in the current-selected account in the UI,
while the sign transaction page uses the account selected by the
transaction, i.e. the sender address. This helps to ensure that the user
knows what account they are signing the transaction with and sending it
from.

Added in the process is a getAccountTotal account selector that resolves
the account total for a specific address.