fix `provider-bridge` indexedDB
Recreate the `dAppPermission` table with the correct primary keys and indexes.

To handle multiple accounts/addresses the primary key for the table has to be a composite key containing both.

The problem is, that Dexie does not support changing primary keys on tables. To work around this limitation I created a multi step migration strategy.