Make redux care about the type selected
Previously the focus was to store the options in redux once
they are loaded from blocknative and just pass them to
components that do some action on them. Here the main
change is that redux will take over some of that logic
and make the type of the fee speed its main concern.