Migrate Redux state from v1 to v2
This migration handles the refactor and rename of
state.ui.currentAccount from a custom `SelectedAccount` type to
state.ui.selectedAccount, a bare `AddressNetwork`.

As this is our first migration of this kind, I've tried to establish
some best practices around handling types in a way that's
forward-compatible. If we can get consensus around this migration
mechanism, it should be extracted from the Main service into its own
module.