Continue the currentAccount refactor
The preferences, provider bridge, and provider services now refer to
the selectedAccount `AddressNetwork` rather than the currentAccount
`string`.

The IndexedDB migration here demonstrates some of the issues we'll have
in the future handling migrations with strict typing. Currently the
service includes two duplicative `Preference` types which might have
been a mistake... but having two means we can more generally type the
one closer to the database to handle migrations, and more narrowly focus
the exported type fore the rest of the code base.

Let's see how it goes.