Fix redux devtools by extending with our BigInt format
Redux devtools hung after a while and was unusable for me. It happened because it did not know how to desierialise our bigint format, even though we had enhancers and whatnot.

The error pointed to [this line of code](https://github.com/kolodny/jsan/blob/master/lib/index.js#L33), which failed when it had to deal with our bigint format. When this was enhanced with our custom serialiser, redux devtools started working.

Tried to fix the issue but it was deep in the stack:
background > `remote-redux-devtools` > `redux-devtools-core` > `remotedev-serialize` > `jsan`. The problem seems to be that `redux-devtools-core` does not propagate our serialiser enhancement here. Can continue going down the rabbit hole to find a good fix but instead changed directions.

It's a simple patch, which fixes our problem. There is this package https://www.npmjs.com/package/patch-package which let's you do that easily. It will apply the patch on every install. (One thing to note is that I DID not install the `postinstall-postinstall` because it seemed that we don't need it and did not want to add dependency unless required)