Move internal provider connections away from ports
Ultimately we may want both port and non-port internal provider
capabilities, but for now it's enough to funnel everything through a
port-esque abstraction that operates entirely in the background script.
This needs to happen if any internal provider traffic comes through the
background script, since a port isn't possible between the background
script and itself.