Rework nonce resolution to occur right before signing
Previously, nonces would be resolved when a transaction was initially
populated. This meant a nonce could be allocated well before a
transaction was due to be broadcast, which increased the probability for
other usage to consume that nonce, resulting in issues with attempting
to replace an existing transaction or to reuse a nonce.

This change populates the nonce immediately before transaction signing.
While in some cases this can still result in a gap between nonce
allocation and broadcast, especially for dApps that use
`eth_signTransaction` instead of `eth_sendTransaction`, it should
reduce the time between nonce allocation and transaction broadcast in
most use cases.