Allow signaling that a nonce is no longer in use
For cases where a nonce was in use but is no longer allocated to a live
transaction---for example, if signing fails for some reason---
ChainService now exposes a releaseEVMTransactionNonce call. The current
behavior is a bit of a sledgehammer, releasing all nonces after the
given transaction's for reuse to prefer errors with replacing
transactions (due to nonce reuse) over transactions that will never mine
(because an earlier nonce hasn't been used).