Parameterize asset amounts for further type specificity
By default, AssetAmount and CompleteAssetAmount both behave as they did
before, with AnyAsset as their asset type. However, they can now be
parametrized with a subtype of AnyAsset, so that an asset amount can be
specified to be only for e.g. SmartContractFungibleAssets, ensuring the
asset itself is a SmartContractFungibleAsset at the type level.