Specify maxBalance for sell assets in swaps
This is done by resolving sell asset amounts, not just assets, and
refining them so that we preserve the SmartContractFungibleAsset type
for the rest of the internals.

This further filters by and sell assets to only
SmartContractFungibleAssets as opposed to all FungibleAssets; for
current purposes, this just means excluding ETH and BTC as possible
assets, since they're unsupported by the backing contracts anyway (and
BTC is a different chain altogether).