Include contract address in SharedAssetInput item key
This allows to distinguish certain assets that have the same symbol but
different contract addresses, for example aToken v1 vs v2 tokens like
aUSDC. Without this change, key interference means the list is not
properly filtered when the search term is updated.

Note that this does not change the fact that two tokens distinguished
solely by their contract address will appear identical to the user in
the UI's asset list.