Only render slide-up menu children when open
Previously, SharedSlideUpMenus would render their children whether or
not they were being displayed. This could lead to a lot of unnecessary
rerendering, but more importantly it made it difficult to use the
component lifecycle for elements that were hidden when the slide-up menu
was hidden, since there was no clear hook to indicate that the element
became visible.

Another option could have been to develop a custom hook, but it
_looks_ like this solution allows standard React component lifecycle
hooks to do the work and doesn't break any expectations in consumers.