Allow more specific types in SharedAssetInput
Before this change, the type was always statically Asset, making it more
difficult to preserve type information in certain cases where all assets
being selected from were of a known Asset subtype.

SharedAssetInput is now parametrized by its asset list, allowing for a
more specific type to be passed out as the selected asset. This means
that if a list of assets entirely consisting of FungibleAssets is used
for the input, for example, the selected asset will be sent to
onAmountChange as a FungibleAsset.