Add support for fully disabling SharedAssetInput
The option existed to disable the dropdown, but in these cases the
amount input was still enabled. In certain scenarios, for example when
an async load of a new value for the input is being performed, it's
desirable to disable the whole input including the amount. The new
isDisabled prop does this.