Do a final asset merge in getCachedAssets
The token list lookup already does merging, but add a final merge pass
with base and custom assets to ensure things are clean. To do this, the
mergeAssets helper ensures that contract address merges also respect the
home network of the given asset, so that two assets with the same
contract address on two separate networks are considered distinct.