Signing only redirects to single asset if available
Previously, SignTransaction would always try to redirect to the
single-asset page—but if the flow did not start at the single-asset
page, the symbol used for this redirect was unset and the result was
an empty page. Additionally, from certain flows (e.g., internal dApps),
going back in history is the correct behavior.

SignTransaction now redirects to the single-asset page if an asset
symbol was passed when the page was navigated to, and otherwise it
navigates back one page in history.