Do not ever clear Ledger state from Redux
Ledger accounts are stored in Redux when shown in the UI,
even if not ultimately imported.
We used to clear them when connecting another Ledger,
to avoid having an ever-growing Redux state.
This commit removes this clear operation,
so that the accounts are kept in Redux forever
and can be used to determine whether a Ledger is available
as the signing methods available for a given address.
The overall size is expected to be manageable
over the entire lifetime of a wallet.