Rework swapPriceJTD to refer to the /price endpoint
This was the JTD for the /prices endpoint, which we no longer use. The
/price (singular) endpoint needs to be used to obtain indicative RFQ-T
liquidity prices for a single pairwise swap; in other words, it needs
to be used instead of `/quote` to get a result close to a firm quote
we would later fetch with `/quote` and `intentOnFilling=true` set.