Rework snackbar hiding to use useDelayContentChange
The snackbar now maintains its content until after the hiding animation
is complete.

For readability, introduce a `shouldHide` variable that is used both as
the `useDelayContentChange` condition and as the class name condition.
This also involved flipping the snackbar element to be tagged with a
`hidden` class instead of an `open` class.

Additionally, pull the dismiss animation timing into a top-level
constant that is used both for `useDelayContentChange` and for the
transition CSS.