Use useDelayContentChange hook in SharedSlideUpMenu
Previously, the slide-up menu would instantly lose its contents on hide,
leading to a slightly disconcerting disappearance of the menu contents
while it slid away. useDelayContentChange lets us delay the hiding of
the contents until after the animation has completed.

In the process, extract the slide transition time to a top-level
constant.

Also included here, `force_hide`/`forceHide`, which was used to avoid
the slide-up menu appearing during inter-page transitions, is now
replaced with an opacity/pointer-events setting directly on the `closed`
class. The transitions for the opacity property are such that the
opacity only goes to 0 after the translateY transition is complete on
hide, and changes immediately to 1 on show.